+++ lib/stream_duration.dart
@@ -1,3 +1,2 @@
-library stream_duration;
-
 export 'src/stream_duration_base.dart';
+export 'src/config/config.dart';