+++ test/sream_duration_test.dart
@@ -1,3 +1,4 @@
+import 'package:stream_duration/src/config/config.dart';
 import 'package:stream_duration/stream_duration.dart';
 import 'package:test/test.dart';
 
@@ -7,16 +8,18 @@ void main() {
     () {
       late StreamDuration streamDuration;
 
-      var expectOutput = [9, 8, 7, 6, 5, 4, 3, 2, 1, 0];
+      final expectOutput = [4, 3, 2, 1, 0];
       var index = 0;
 
       setUp(
         () {
           streamDuration = StreamDuration(
-            Duration(seconds: 10),
-            onDone: () {
-              print('Stream Donde & disposed');
-            },
+            config: StreamDurationConfig(
+              countDownConfig: const CountDownConfig(
+                duration: Duration(seconds: 5),
+              ),
+              onDone: () => print('Stream Done'),
+            ),
           );
         },
       );
@@ -38,19 +41,57 @@ void main() {
 
   test('Change duration test', () {
     final streamDuration = StreamDuration(
-      Duration(seconds: 10),
-      onDone: () {
-        print('Stream Donde & disposed');
-      },
+      config: StreamDurationConfig(
+        countDownConfig: const CountDownConfig(
+          duration: Duration(seconds: 10),
+        ),
+        onDone: () => print('Stream Donde & disposed'),
+      ),
     );
 
     // Expect initalize duration is match
     expect(streamDuration.remainingDuration.inSeconds, 10);
 
     // Change duration
-    streamDuration.change(Duration(seconds: 20));
+    streamDuration.change(const Duration(seconds: 20));
 
     // Expect change duration match
     expect(streamDuration.remainingDuration.inSeconds, 20);
+  });
+
+  group('countUp test', () {
+    late StreamDuration streamDuration;
+
+    final expectOutput = [1, 2, 3, 4];
+    var index = 0;
+
+    setUp(
+      () {
+        streamDuration = StreamDuration(
+          config: StreamDurationConfig(
+            isCountUp: true,
+            countUpConfig: const CountUpConfig(
+              initialDuration: Duration.zero,
+              maxDuration: const Duration(seconds: 5),
+            ),
+            onDone: () => print('Stream Done'),
+          ),
+        );
+      },
+    );
+
+    test(
+      'Return Stream match',
+      () {
+        streamDuration.durationLeft.listen(
+          expectAsync1(
+            (event) {
+              expect(event.inSeconds, expectOutput[index]);
+              index++;
+            },
+          ),
+        );
+      },
+    );
   });
 }